renamed package to teacher portal and updated all files including it
