--- teacher/src/main/java/com/projectsuccess/teacher/Teacher.java
@@ -1,11 +1,0 @@
-package com.project-success.teacher;
-
-@Setter
-@Getter
-public class Teacher {
-    public int teacherId;
-    private String firstName;
-    private String lastName;
-    private List<Course> courses;
-
-}
