+++ teacher/src/main/java/com/projectsuccess/teacherportal/TeacherController.java
@@ -1,10 +1,10 @@
-package com.example.handlingformsubmission;
+package com.projectsuccess.teacherportal;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class GreetingController {
+public class TeacherController {
 
     @GetMapping("/assignment")
         public String getAssignmentDetails(@RequestParam int assignmentId) {
