+++ teacher/src/main/java/com/projectsuccess/teacherportal/TeacherPortal.java
@@ -1,4 +1,4 @@
-package com.projectsuccess.teacher;
+package com.projectsuccess.teacherportal;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
